+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentSheet.kt
@@ -58,6 +58,10 @@ class PaymentSheet internal constructor(
      * @param activity The Activity that is presenting [PaymentSheet].
      * @param callback Called with the result of the payment after [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith("PaymentSheet.Builder(callback).build(activity)")
+    )
     constructor(
         activity: ComponentActivity,
         callback: PaymentSheetResultCallback
@@ -73,6 +77,14 @@ class PaymentSheet internal constructor(
      * @param externalPaymentMethodConfirmHandler Called when a user confirms payment with an external payment method.
      * @param callback Called with the result of the payment after [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(callback)" +
+                ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                ".build(activity)"
+        )
+    )
     constructor(
         activity: ComponentActivity,
         externalPaymentMethodConfirmHandler: ExternalPaymentMethodConfirmHandler,
@@ -96,6 +108,14 @@ class PaymentSheet internal constructor(
      * @param paymentResultCallback Called with the result of the payment or setup after
      * [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(paymentResultCallback)" +
+                ".createIntentCallback(createIntentCallback)" +
+                ".build(activity)"
+        )
+    )
     constructor(
         activity: ComponentActivity,
         createIntentCallback: CreateIntentCallback,
@@ -121,6 +141,15 @@ class PaymentSheet internal constructor(
      * @param paymentResultCallback Called with the result of the payment or setup after
      * [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(paymentResultCallback)" +
+                ".createIntentCallback(createIntentCallback)" +
+                ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                ".build(activity)"
+        )
+    )
     constructor(
         activity: ComponentActivity,
         createIntentCallback: CreateIntentCallback,
@@ -143,6 +172,10 @@ class PaymentSheet internal constructor(
      * @param fragment the Fragment that is presenting the payment sheet.
      * @param callback called with the result of the payment after the payment sheet is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith("PaymentSheet.Builder(callback).build(fragment)")
+    )
     constructor(
         fragment: Fragment,
         callback: PaymentSheetResultCallback
@@ -158,6 +191,14 @@ class PaymentSheet internal constructor(
      * @param externalPaymentMethodConfirmHandler Called when a user confirms payment with an external payment method.
      * @param callback called with the result of the payment after the payment sheet is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(callback)" +
+                ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                ".build(fragment)"
+        )
+    )
     constructor(
         fragment: Fragment,
         externalPaymentMethodConfirmHandler: ExternalPaymentMethodConfirmHandler,
@@ -181,6 +222,14 @@ class PaymentSheet internal constructor(
      * @param paymentResultCallback Called with the result of the payment or setup after
      * [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(paymentResultCallback)" +
+                ".createIntentCallback(createIntentCallback)" +
+                ".build(fragment)"
+        )
+    )
     constructor(
         fragment: Fragment,
         createIntentCallback: CreateIntentCallback,
@@ -206,6 +255,15 @@ class PaymentSheet internal constructor(
      * @param paymentResultCallback Called with the result of the payment or setup after
      * [PaymentSheet] is dismissed.
      */
+    @Deprecated(
+        message = "This will be removed in a future release.",
+        replaceWith = ReplaceWith(
+            "PaymentSheet.Builder(paymentResultCallback)" +
+                ".createIntentCallback(createIntentCallback)" +
+                ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                ".build(fragment)"
+        )
+    )
     constructor(
         fragment: Fragment,
         createIntentCallback: CreateIntentCallback,
@@ -2720,6 +2778,12 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback).build(activity)"
+                )
+            )
             fun create(
                 activity: ComponentActivity,
                 paymentOptionCallback: PaymentOptionCallback,
@@ -2747,6 +2811,14 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                        ".build(activity)"
+                )
+            )
             fun create(
                 activity: ComponentActivity,
                 externalPaymentMethodConfirmHandler: ExternalPaymentMethodConfirmHandler,
@@ -2777,6 +2849,14 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".createIntentCallback(createIntentCallback)" +
+                        ".build(activity)"
+                )
+            )
             fun create(
                 activity: ComponentActivity,
                 paymentOptionCallback: PaymentOptionCallback,
@@ -2811,6 +2891,15 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".createIntentCallback(createIntentCallback)" +
+                        ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                        ".build(activity)"
+                )
+            )
             fun create(
                 activity: ComponentActivity,
                 paymentOptionCallback: PaymentOptionCallback,
@@ -2840,6 +2929,12 @@ class PaymentSheet internal constructor(
              * @param paymentResultCallback called when a [PaymentSheetResult] is available.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback).build(fragment)"
+                )
+            )
             fun create(
                 fragment: Fragment,
                 paymentOptionCallback: PaymentOptionCallback,
@@ -2865,6 +2960,14 @@ class PaymentSheet internal constructor(
              * @param paymentResultCallback called when a [PaymentSheetResult] is available.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                        ".build(fragment)"
+                )
+            )
             fun create(
                 fragment: Fragment,
                 externalPaymentMethodConfirmHandler: ExternalPaymentMethodConfirmHandler,
@@ -2895,6 +2998,14 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".createIntentCallback(createIntentCallback)" +
+                        ".build(fragment)"
+                )
+            )
             fun create(
                 fragment: Fragment,
                 paymentOptionCallback: PaymentOptionCallback,
@@ -2929,6 +3040,15 @@ class PaymentSheet internal constructor(
              * [PaymentSheet] is dismissed.
              */
             @JvmStatic
+            @Deprecated(
+                message = "This will be removed in a future release.",
+                replaceWith = ReplaceWith(
+                    "FlowController.Builder(paymentResultCallback, paymentOptionCallback)" +
+                        ".createIntentCallback(createIntentCallback)" +
+                        ".externalPaymentMethodConfirmHandler(externalPaymentMethodConfirmHandler)" +
+                        ".build(fragment)"
+                )
+            )
             fun create(
                 fragment: Fragment,
                 paymentOptionCallback: PaymentOptionCallback,